requirements.txt
flask
tensorflow
numpy
pillow
werkzeug
torch
torchvision
